all done

+++ FPH.DataBase/Repositories/PaymentTypeRepository.cs
@@ -3,6 +3,7 @@ using FPH.DataBase.Abstractions;
 using FPH.DataBase.Context;
 using Microsoft.EntityFrameworkCore;
 
+
 namespace FPH.DataBase.Repositories
 {
     public class PaymentTypeRepository : IPaymentTypeRepository
@@ -13,33 +14,37 @@ namespace FPH.DataBase.Repositories
         {
             _context = context;
         }
-
-        public async Task<PaymentTypeEntity> GetByIdAsync(int id)
-        {
-            return await _context.PaymentTypes.FindAsync(id);
-        }
-
-        public async Task<IEnumerable<PaymentTypeEntity>> GetAllAsync()
+        public async Task<IEnumerable<PaymentTypeEntity>> GetAllPaymentTypesAsync()
         {
             return await _context.PaymentTypes.ToListAsync();
         }
 
-        public async Task AddAsync(PaymentTypeEntity paymentType)
+        public async Task<PaymentTypeEntity> GetPaymentTypeByIdAsync(int id)
+        {
+            return await _context.PaymentTypes.FirstOrDefaultAsync(x => x.Id == id);
+        }
+
+        public async Task AddPaymentTypeAsync(PaymentTypeEntity paymentType)
         {
             _context.PaymentTypes.Add(paymentType);
             await _context.SaveChangesAsync();
         }
 
-        public async Task UpdateAsync(PaymentTypeEntity paymentType)
+        public async Task UpdatePaymentTypeAsync(PaymentTypeEntity paymentType)
         {
             _context.Entry(paymentType).State = EntityState.Modified;
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(PaymentTypeEntity paymentType)
+        public async Task DeletePaymentTypeAsync(int id)
         {
-            _context.PaymentTypes.Remove(paymentType);
-            await _context.SaveChangesAsync();
+            var paymentType = await _context.PaymentTypes.FindAsync(id);
+            if (paymentType != null)
+            {
+                _context.PaymentTypes.Remove(paymentType);
+                await _context.SaveChangesAsync();
+            }
         }
+
     }
 }
